pyproject.toml
[project]
name = "chatbot_finance"
version = "0.1.0"
description = "Financial chatbot project from BlentAI platform."
requires-python = ">=3.12"
dependencies = [
    "bert-score>=0.3.13",
    "chromadb>=1.0.13",
    "cryptography>=3.1",
    "dotenv>=0.9.9",
    "gradio>=5.34.2",
    "huggingface-hub>=0.33.0",
    "ipywidgets>=8.1.7",
    "langchain>=0.3.26",
    "langchain-chroma>=0.2.4",
    "langchain-community>=0.3.26",
    "langchain-huggingface>=0.3.0",
    "llama-cpp-python>=0.3.9",
    "notebook>=7.4.3",
    "openai>=1.91.0",
    "pypdf>=5.6.1",
    "sacremoses>=0.1.1",
    "sentence-transformers>=4.1.0",
]

[tool.uv]
dev-dependencies = [
    "black>=25.1.0",
    "ipykernel>=6.29.5",
    "pylint>=3.3.6",
    "pytest>=8.3.5",
]

[tool.pylint.'MAIN']
max-line-length = 100
disable = ["invalid-name", "broad-exception-caught"]

[build-system]
requires = ["setuptools>=58", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["chatbot_finance"]
exclude = ["data", "notebooks", "scripts"] 
